requirements.txt
pandas
numpy
mysql-connector-python
openai
matplotlib
xlsxwriter
schedule
